refactor: add correct name validation
#2318 #2926
