+++ visualization/app/codeCharta/ui/scenarioDropDown/addCustomScenario/addCustomScenario.component.ts
@@ -1,16 +1,18 @@
 import { Component, Inject } from "@angular/core"
+import { FormControl } from "@angular/forms"
 import { ThreeCameraServiceToken, ThreeOrbitControlsServiceToken } from "../../../services/ajs-upgraded-providers"
 import { State } from "../../../state/angular-redux/state"
 import { ScenarioHelper, ScenarioMetricProperty } from "../../../util/scenarioHelper"
 import { ThreeCameraService } from "../../codeMap/threeViewer/threeCameraService"
 import { ThreeOrbitControlsService } from "../../codeMap/threeViewer/threeOrbitControlsService"
-import { getInitialScenarioMetricProperties } from "./getInitialScenarioMetricProperties"
+import { customScenarioNameValidator } from "./utils/customScenarioName.validator"
+import { getInitialScenarioMetricProperties } from "./utils/getInitialScenarioMetricProperties"
 
 @Component({
 	template: require("./addCustomScenario.component.html")
 })
 export class AddCustomScenarioComponent {
-	scenarioName = ""
+	scenarioName = new FormControl("", [customScenarioNameValidator()])
 	scenarioNameErrorField: string | null = "Scenario name is required"
 	scenarioContent: ScenarioMetricProperty[]
 	areAnyScenarioMetricPropertiesSelected = true
@@ -26,25 +28,12 @@ export class AddCustomScenarioComponent {
 		})
 	}
 
-	validateScenarioName(event: Event) {
-		const newScenarioName = (event.target as HTMLInputElement).value
-		if (newScenarioName.length === 0) {
-			this.scenarioNameErrorField = "Scenario name is required"
-		} else if (ScenarioHelper.isScenarioExisting(newScenarioName)) {
-			// todo this error isn't displayed
-			this.scenarioNameErrorField = "A Scenario with this name already exists"
-		} else {
-			this.scenarioNameErrorField = null
-		}
-		// console.log(this.scenarioNameErrorField)
-	}
-
 	handleScenarioMetricPropertySelectionChange(scenarioMetricProperty: ScenarioMetricProperty) {
 		scenarioMetricProperty.isSelected = !scenarioMetricProperty.isSelected
 		this.areAnyScenarioMetricPropertiesSelected = this.scenarioContent.some(property => property.isSelected)
 	}
 
 	addCustomScenario() {
-		ScenarioHelper.addScenario(this.scenarioName, this.scenarioContent)
+		ScenarioHelper.addScenario(this.scenarioName.value, this.scenarioContent)
 	}
 }
