+++ visualization/app/codeCharta/ui/scenarioDropDown/addCustomScenario/addCustomScenario.module.ts
@@ -1,11 +1,11 @@
 import { CommonModule } from "@angular/common"
 import { NgModule } from "@angular/core"
-import { FormsModule } from "@angular/forms"
+import { FormsModule, ReactiveFormsModule } from "@angular/forms"
 import { MaterialModule } from "../../../../material/material.module"
 import { AddCustomScenarioComponent } from "./addCustomScenario.component"
 
 @NgModule({
-	imports: [CommonModule, FormsModule, MaterialModule],
+	imports: [CommonModule, FormsModule, ReactiveFormsModule, MaterialModule],
 	declarations: [AddCustomScenarioComponent],
 	entryComponents: [AddCustomScenarioComponent]
 })
