+++ visualization/app/codeCharta/ui/scenarioDropDown/addCustomScenario/utils/customScenarioName.validator.ts
@@ -1,0 +1,14 @@
+import { ValidatorFn, AbstractControl, ValidationErrors } from "@angular/forms"
+import { ScenarioHelper } from "../../../../util/scenarioHelper"
+
+export function customScenarioNameValidator(): ValidatorFn {
+	return (control: AbstractControl): ValidationErrors | null => {
+		if (control.value.length === 0) {
+			return { Error: "Scenario name is required" }
+		}
+		if (ScenarioHelper.isScenarioExisting(control.value)) {
+			return { Error: "A Scenario with this name already exists" }
+		}
+		return null
+	}
+}
